refactor bbclib.py   Every class in bbclib_core.py is defined in an individual file
