+++ bbc1/core/data_handler.py
@@ -20,11 +20,10 @@ import binascii
 import os
 import sys
 sys.path.extend(["../../", os.path.abspath(os.path.dirname(__file__))])
-from bbc1.core import bbclib_core, bbclib, bbclib_wire
+from bbc1.core import bbclib
+import bbc1.core.libs.bbclib_config as bbclib_config
 from bbc1.core.message_key_types import to_2byte, PayloadType, KeyType, InfraMessageCategory
 from bbc1.core import logger
-
-DEFAULT_BBC_FORMAT = bbclib_wire.BBcFormat.FORMAT_PLAIN
 
 
 transaction_tbl_definition = [
@@ -77,7 +76,7 @@ class DataHandler:
         self.stats = networking.core.stats
         self.logger = logger.get_logger(key="data_handler", level=loglevel, logname=logname)
         self.domain_id = domain_id
-        self.domain_id_str = bbclib_core.convert_id_to_string(domain_id)
+        self.domain_id_str = bbclib.convert_id_to_string(domain_id)
         self.config = config
         self.working_dir = workingdir
         self.storage_root = os.path.join(self.working_dir, self.domain_id_str)
@@ -213,7 +212,7 @@ class DataHandler:
                 info.append((txobj.transaction_id, pt.transaction_id))  # (base, point_to)
         return info
 
-    def insert_transaction(self, txdata, txobj=None, fmt_type=DEFAULT_BBC_FORMAT, asset_files=None, no_replication=False):
+    def insert_transaction(self, txdata, txobj=None, fmt_type=bbclib_config.DEFAULT_BBC_FORMAT, asset_files=None, no_replication=False):
         """Insert transaction data and its asset files
 
         Either txdata or txobj must be given to insert the transaction.
@@ -255,7 +254,7 @@ class DataHandler:
 
         return asset_group_ids
 
-    def _insert_transaction_into_a_db(self, db_num, txobj, fmt_type=DEFAULT_BBC_FORMAT):
+    def _insert_transaction_into_a_db(self, db_num, txobj, fmt_type=bbclib_config.DEFAULT_BBC_FORMAT):
         """Insert transaction data into the transaction table of the specified DB
 
         Args:
