+++ bbc1/core/libs/bbclib_msgtype.py
@@ -1,0 +1,82 @@
+# -*- coding: utf-8 -*-
+"""
+Copyright (c) 2018 beyond-blockchain.org.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+"""
+
+
+class MsgType:
+    """Message types for between core node and client"""
+    REQUEST_SETUP_DOMAIN = 0
+    RESPONSE_SETUP_DOMAIN = 1
+    REQUEST_SET_STATIC_NODE = 4
+    RESPONSE_SET_STATIC_NODE = 5
+    REQUEST_GET_CONFIG = 8
+    RESPONSE_GET_CONFIG = 9
+    REQUEST_MANIP_LEDGER_SUBSYS = 10
+    RESPONSE_MANIP_LEDGER_SUBSYS = 11
+    DOMAIN_PING = 12
+    REQUEST_GET_DOMAINLIST = 13
+    RESPONSE_GET_DOMAINLIST = 14
+    REQUEST_INSERT_NOTIFICATION = 15
+    CANCEL_INSERT_NOTIFICATION = 16
+    REQUEST_GET_STATS = 17
+    RESPONSE_GET_STATS = 18
+    NOTIFY_DOMAIN_KEY_UPDATE = 19
+    REQUEST_GET_NEIGHBORLIST = 21
+    RESPONSE_GET_NEIGHBORLIST = 22
+    REQUEST_GET_USERS = 23
+    RESPONSE_GET_USERS = 24
+    REQUEST_GET_FORWARDING_LIST = 25
+    RESPONSE_GET_FORWARDING_LIST = 26
+    REQUEST_GET_NODEID = 27
+    RESPONSE_GET_NODEID = 28
+    REQUEST_GET_NOTIFICATION_LIST = 29
+    RESPONSE_GET_NOTIFICATION_LIST = 30
+    REQUEST_CLOSE_DOMAIN = 31
+    RESPONSE_CLOSE_DOMAIN = 32
+    REQUEST_ECDH_KEY_EXCHANGE = 33
+    RESPONSE_ECDH_KEY_EXCHANGE = 34
+
+    REGISTER = 64
+    UNREGISTER = 65
+    MESSAGE = 66
+
+    REQUEST_GATHER_SIGNATURE = 67
+    RESPONSE_GATHER_SIGNATURE = 68
+    REQUEST_SIGNATURE = 69
+    RESPONSE_SIGNATURE = 70
+    REQUEST_INSERT = 71
+    RESPONSE_INSERT = 72
+    NOTIFY_INSERTED = 73
+    NOTIFY_CROSS_REF = 74
+
+    REQUEST_SEARCH_TRANSACTION = 82
+    RESPONSE_SEARCH_TRANSACTION = 83
+    REQUEST_SEARCH_WITH_CONDITIONS = 86
+    RESPONSE_SEARCH_WITH_CONDITIONS = 87
+    REQUEST_TRAVERSE_TRANSACTIONS = 88
+    RESPONSE_TRAVERSE_TRANSACTIONS = 89
+    REQUEST_CROSS_REF_VERIFY = 90
+    RESPONSE_CROSS_REF_VERIFY = 91
+    REQUEST_CROSS_REF_LIST = 92
+    RESPONSE_CROSS_REF_LIST = 93
+    REQUEST_REPAIR = 94
+    REQUEST_COUNT_TRANSACTIONS = 95
+    RESPONSE_COUNT_TRANSACTIONS = 95
+
+    REQUEST_REGISTER_HASH_IN_SUBSYS = 128
+    RESPONSE_REGISTER_HASH_IN_SUBSYS = 129
+    REQUEST_VERIFY_HASH_IN_SUBSYS = 130
+    RESPONSE_VERIFY_HASH_IN_SUBSYS = 131
