+++ bbc1/core/libs/bbclib_wire.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-Copyright (c) 2017 beyond-blockchain.org.
+Copyright (c) 2018 beyond-blockchain.org.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
