--- docs/api/bbc1.core.bbclib.rst
@@ -1,11 +1,0 @@
-bbc1.core.bbclib module
-=======================
-
-.. automodule:: bbc1.core.bbclib
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-    .. autosummary::
-        :toctree: _gen
-
