+++ prepare.sh
@@ -8,7 +8,7 @@ bash prepare.sh
 
 cd lib
 if [ -f libbbcsig.dylib ]; then
-  cp libbbcsig.dylib ../../bbc1/core/
+  cp libbbcsig.dylib ../../bbc1/core/libs/
 elif [ -f libbbcsig.so ]; then
-  cp libbbcsig.so ../../bbc1/core/
+  cp libbbcsig.so ../../bbc1/core/libs/
 fi
